Lotsa stuff, primarily github-release
* lucli github-release, borrowed from staticli, to release new versions of lucli to github
* Switched to Github for releases. Sorry, BinTray
* Terraform now supports arbitrary versions, defaulting to latest
* Other misc tweaks

Basically stuff that's been in my repo for the past few weeks that I've not bothered to commit yet

+++ cmd/jess.go
@@ -11,7 +11,7 @@ func init() {
 
 	command := cli.NewCommand("jess [params]")
 	command.SetShort("Very basic generic Jessie Frazelle style GUI Container Runner")
-	command.SetShort(`Runs containers as simply as possible.
+	command.SetLong(`Runs containers as simply as possible.
 
 This is basically the equivalent of:
